streamlit_app/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
xgboost>=1.5.0
catboost>=1.0.0
joblib>=1.0.0
streamlit>=1.12.0
plotly>=5.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# NEW: RAG Integration Dependencies
httpx>=0.27.0
requests>=2.32.0
python-dotenv>=1.0.0